+++ src/st7789/ST7789_TFT_LCD_RVL.cpp
@@ -112,7 +112,9 @@ void ST7789_TFT ::TFTPowerDown(void)
 {
 	TFTenableDisplay(false);
 	DisplayRVL_DC_SetLow;
-	DisplayRVL_RST_SetLow;
+	if (_resetPinOn){
+		DisplayRVL_RST_SetLow;
+	}
 
 	if (_hardwareSPI == false)
 	{
@@ -134,26 +136,32 @@ void ST7789_TFT ::TFTPowerDown(void)
 	@brief: Method for Hardware Reset pin control
 */
 void ST7789_TFT ::TFTResetPIN() {
-	DisplayRVL_RST_SetDigitalOutput;
-	DisplayRVL_RST_SetHigh;
-	const uint8_t TFT_RESET_DELAY = 10; /**< Reset delay in mS*/
-	delayMilliSecRVL (TFT_RESET_DELAY);
-	DisplayRVL_RST_SetLow;
-	delayMilliSecRVL (TFT_RESET_DELAY);
-	DisplayRVL_RST_SetHigh;
-	delayMilliSecRVL (TFT_RESET_DELAY);
+	if (_resetPinOn == true)
+	{
+		DisplayRVL_RST_SetDigitalOutput;
+		DisplayRVL_RST_SetHigh;
+		const uint8_t TFT_RESET_DELAY = 10; /**< Reset delay in mS*/
+		delayMilliSecRVL (TFT_RESET_DELAY);
+		DisplayRVL_RST_SetLow;
+		delayMilliSecRVL (TFT_RESET_DELAY);
+		DisplayRVL_RST_SetHigh;
+		delayMilliSecRVL (TFT_RESET_DELAY);
+	}else{
+		std::cout << "Error:TFTResetPIN: User has called this function, yet reset is set to software mode " << std::endl;
+	}
 }
 
 /*!
 	@brief  sets up TFT GPIO for Hardware SPi
 	@param rst reset GPIO
 	@param dc data or command GPIO.
+	@details for software reset pass -1 to rst
 	@note overloaded 2 off, 1 for HW SPI , 1 for SW SPI 
 */
 void ST7789_TFT ::TFTSetupGPIO(int8_t rst, int8_t dc)
 {
 	_hardwareSPI = true;
-	_DisplayRVL_RST= rst;
+	TFTSetupResetPin(rst);
 	_DisplayRVL_DC = dc;
 }
 
@@ -164,6 +172,7 @@ void ST7789_TFT ::TFTSetupGPIO(int8_t rst, int8_t dc)
 	@param cs chip select GPIO
 	@param sclk Data clock GPIO
 	@param din Data to TFT GPIO
+	@details for software reset pass -1 to rst
 	@note overloaded 2 off, 1 for HW SPI , 1 for SW SPI 
 */
 void ST7789_TFT ::TFTSetupGPIO(int8_t rst, int8_t dc, int8_t cs, int8_t sclk, int8_t din)
@@ -172,7 +181,7 @@ void ST7789_TFT ::TFTSetupGPIO(int8_t rst, int8_t dc, int8_t cs, int8_t sclk, in
 	_DisplayRVL_CS = cs;
 	_DisplayRVL_SDATA = din;
 	_DisplayRVL_SCLK = sclk;
-	_DisplayRVL_RST= rst;
+	TFTSetupResetPin(rst);
 	_DisplayRVL_DC = dc;
 }
 
@@ -182,7 +191,10 @@ void ST7789_TFT ::TFTSetupGPIO(int8_t rst, int8_t dc, int8_t cs, int8_t sclk, in
 	@return rvlDisplay_SPIbeginFail error if bcm2835_spi_begin or bcm2835_aux_spi_begin has failed
 */
 rvlDisplay_Return_Codes_e ST7789_TFT::TFTST7789Initialize() {
-	TFTResetPIN();
+	if (_resetPinOn == true)
+	{
+		TFTResetPIN();
+	}
 	DisplayRVL_DC_SetDigitalOutput;
 	DisplayRVL_DC_SetLow;
 if (_hardwareSPI == false)
@@ -459,16 +471,19 @@ void  ST7789_TFT::HighFreqDelaySet(uint16_t CommDelay){_HighFreqDelay = CommDela
 	@note virtual function overloads graphics library
  */
 void ST7789_TFT::setAddrWindow(uint16_t x0, uint16_t y0, uint16_t x1, uint16_t y1) {
-	uint8_t x0Higher = (x0 >> 8) + _XStart;
-	uint8_t x0Lower  = (x0 &  0xFF) + _XStart;
-	uint8_t y0Higher = (y0 >> 8) + _YStart;
-	uint8_t y0Lower  = (y0 &  0xFF) + _YStart;
-	uint8_t x1Higher = (x1 >> 8) + _XStart;
-	uint8_t x1Lower  = (x1 &  0xFF) + _XStart;
-	uint8_t y1Higher = (y1 >> 8) + _YStart;
-	uint8_t y1Lower  = (y1 &  0xFF) + _YStart;
-	uint8_t seqCASET[]    {x0Higher ,x0Lower,x1Higher,x1Lower};
-	uint8_t seqRASET[]    {y0Higher,y0Lower,y1Higher,y1Lower};
+	uint16_t x0_ = x0 + _XStart;
+	uint16_t x1_ = x1 + _XStart;
+	uint16_t y0_ = y0 + _YStart;
+	uint16_t y1_ = y1 + _YStart;
+	uint8_t seqCASET[] = {
+		(uint8_t)(x0_ >> 8), (uint8_t)(x0_ & 0xFF),
+		(uint8_t)(x1_ >> 8), (uint8_t)(x1_ & 0xFF)
+	};
+	uint8_t seqRASET[] = {
+		(uint8_t)(y0_ >> 8), (uint8_t)(y0_ & 0xFF),
+		(uint8_t)(y1_ >> 8), (uint8_t)(y1_ & 0xFF)
+	};
+
 	writeCommand(ST7789_CASET); //Column address set
 	spiWriteDataBuffer(seqCASET, sizeof(seqCASET));
 	writeCommand(ST7789_RASET); //Row address set
@@ -482,9 +497,9 @@ void ST7789_TFT::setAddrWindow(uint16_t x0, uint16_t y0, uint16_t x1, uint16_t y
 */
 void ST7789_TFT::cmd89(void)
 {
-	uint8_t CASETsequence[] {0x00, 0x00, uint8_t(_widthStartTFT  >> 8),  uint8_t(_widthStartTFT  & 0xFF)};
-	uint8_t RASETsequence[] {0x00, 0x00, uint8_t(_heightStartTFT >> 8) , uint8_t(_heightStartTFT & 0XFF)};
-	
+	uint8_t CASETsequence[] {0x00, 0x00, 0,  240};
+	uint8_t RASETsequence[] {0x00, 0x00, 320 >> 8 , 320 & 0xFF};
+
 	writeCommand(ST7789_SWRESET);
 	delayMilliSecRVL (150);
 	writeCommand(ST7789_SLPOUT);
@@ -528,6 +543,22 @@ void ST7789_TFT::AdjustWidthHeight() {
 	// 1.47", 1.69, 1.9", 2.0" displays (centered)
 		_rowstart = _rowstart2 = (int)((320 - _height) / 2);
 		_colstart = _colstart2 = (int)((240 - _width) / 2);
+	}
+}
+
+/*!
+	@brief sets up TFT GPIO reset pin
+	@param rst reset GPIO
+	@details if rst value is -1 use software reset, else use Hardware reset.
+*/
+void ST7789_TFT::TFTSetupResetPin(int8_t rst)
+{
+	if(rst != -1)
+	{
+		_DisplayRVL_RST= rst;
+		_resetPinOn = true;
+	}else{
+		_resetPinOn  = false;
 	}
 }