+++ extra/doc/CHANGELOG.md
@@ -6,3 +6,6 @@ Version control history
 	* A C++ Library to connect St7789 electronic displays to Raspberry Pi single board computers.
 * version 1.0.2 11-2024
 	* Added support for hardware SPI 1 (AUX) port
+* version 1.1.0 06-2025
+	* Added Support for optional reset pin.
+	* Updates to functions cmd89() and setAddrWindow().